Initial work towards a port for 1.16
Hello, I've tried the best I can to port your mod to 1.16.1 from 1.14 and have fixed up a bunch of errors caused by the changing in mappings. I would've ported the whole thing to 1.16.1 if I had the time and knowledge to do so, but as I don't this is the best I can do.

Access Transformers, coremodding js files and overrides have been left untouched.

Hope this helps! :)